move metrics_gauge_task to package init
